Refactor(InputValidator): 차 이름 입력에 대한 책임을 CarValidator로 분리
